// src/shared/api/generated/model/createRoutineRequestDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Yoclick API
 * API multi-centro de reservas en marca blanca. Contrato canónico para la app.
 * OpenAPI spec version: 0.1.0
 */
import type { CreateRoutineRequestDtoAssignTo } from './createRoutineRequestDtoAssignTo.ts';
import type { CreateRoutineRequestDtoItemsItem } from './createRoutineRequestDtoItemsItem.ts';

export interface CreateRoutineRequestDto {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /**
     * @maxLength 500
     * @nullable
     */
  note?: string | null;
  /**
     * @minItems 1
     * @maxItems 30
     */
  items: CreateRoutineRequestDtoItemsItem[];
  assignTo?: CreateRoutineRequestDtoAssignTo;
}